adding ppl ready
