+++ src/components/RightPanel.jsx
@@ -1,11 +1,8 @@
 import React from "react";
-import { useState, useEffect } from "react";
-import { Col } from "react-bootstrap";
 import MessageSend from "./MessageSend";
 import MessagesWindow from "./MessagesWindow";
-import axios from "axios";
-import ChatList from "./ChatList";
 import GroupSettings from "./GroupSettings";
+import { Link } from "react-router-dom";
 
 export default function (props) {
   const showMessWindow = props.rightPanelMode === "messWin";
@@ -16,32 +13,20 @@ export default function (props) {
     <div className="second">
       <div className="rightPanelUpper">
         <label> User: {props.selectedUser?.queueName}</label>
-        <button
-          onClick={() => {
-            props.setRightPanelMode("messSett");
-          }}
-        >
-          Settings
-        </button>
+        <Link to={"/index/settings"}>Settings</Link>
       </div>
 
       <div className="rightPanelMid">
         {showMessWindow ? (
           <MessagesWindow
             account={props.account}
-            allUsers={props.selectedChat.users}
             messages={props.messages}
             refetchMess={props.refetchMess}
             refetchAcc={props.refetchAcc}
           />
         ) : null}
         {showSettWindow ? (
-          <GroupSettings
-            selectedChat={props.selectedChat}
-            account={props.account}
-            allUsers={props.selectedChat.users}
-            messages={props.messages}
-          />
+          <GroupSettings account={props.account} messages={props.messages} />
         ) : null}
 
         {showWelcomeWindow ? <h1>WELCOME</h1> : null}
@@ -51,7 +36,6 @@ export default function (props) {
         <MessageSend
           refetchMess={props.refetchMess}
           refetchAcc={props.refetchAcc}
-          queueId={props.selectedChat?.queueId}
           senderId={props.account?.userId}
         />
       </div>
